+++ harmonySubPlatform.js
@@ -1,4 +1,4 @@
-var Service, Characteristic, Accessory;
+var Service, Characteristic, Accessory, AccessControlManagement, AccessControlEvent;
 const HarmonyBase = require('./harmonyBase').HarmonyBase;
 const HarmonyConst = require('./harmonyConst');
 const HarmonyTools = require('./harmonyTools.js');
@@ -13,6 +13,8 @@ function HarmonySubPlatform(log, config, api, mainPlatform) {
   Service = api.hap.Service;
   Characteristic = api.hap.Characteristic;
   Accessory = api.hap.Accessory;
+  AccessControlManagement = api.hap.AccessControlManagement;
+  AccessControlEvent = api.hap.AccessControlEvent;
 
   this.api = api;
   this.mainPlatform = mainPlatform;
@@ -278,6 +280,77 @@ HarmonySubPlatform.prototype = {
       else this.configureMainActivity(myHarmonyAccessory, defaultActivity);
     }
 
+    ////
+
+    //acces control
+    let subType = this.name + ' AccessControlService';
+    this.accessControlService = myHarmonyAccessory.getServiceByUUIDAndSubType(this.name, subType);
+    var accessControl;
+
+    if (!this.accessControlService) {
+      accessControl = new AccessControlManagement(true);
+      this.accessControlService = accessControl.getService();
+    } else {
+      accessControl = new AccessControlManagement(true, this.accessControlService);
+    }
+
+    this._confirmedServices.push(this.accessControlService);
+    myHarmonyAccessory.addService(this.accessControlService);
+
+    accessControl.on(AccessControlEvent.ACCESS_LEVEL_UPDATED, (level) => {
+      this.log.debug(
+        '(' +
+          this.name +
+          ')' +
+          'INFO - New access control level of ' +
+          level +
+          '/' +
+          myHarmonyAccessory.displayName
+      );
+    });
+    accessControl.on(AccessControlEvent.PASSWORD_SETTING_UPDATED, (password, passwordRequired) => {
+      if (passwordRequired) {
+        this.log.debug(
+          '(' +
+            this.name +
+            ')' +
+            'INFO - New access control - Required password is: ' +
+            password +
+            '/' +
+            myHarmonyAccessory.displayName
+        );
+      } else {
+        this.log.debug(
+          '(' +
+            this.name +
+            ')' +
+            'INFO - access control - No password set! ' +
+            '/' +
+            myHarmonyAccessory.displayName
+        );
+      }
+    });
+
+    /*
+        const myController = new RemoteController();
+        // ----------------- for siri support -----------------
+       // import { GStreamerAudioProducer, GStreamerOptions } from "./gstreamer-audioProducer";
+// CHANGE this to enable siri support. Read docs in 'gstreamer-audioProducer.ts' for necessary package dependencies
+//const siriSupport = false;
+//const gstreamerOptions: Partial<GStreamerOptions> = { // any configuration regarding the producer can be made here
+// ----------------------------------------------------
+        //new RemoteController(GStreamerAudioProducer, gstreamerOptions)
+
+        myHarmonyAccessory.configureController(myController)
+
+        this._confirmedServices.push(myController.targetControlManagementService);
+        this._confirmedServices.push(myController.targetControlService);
+        //TO ENABLE SIRI this._confirmedServices.push(myController.audioStreamManagementService);
+        this.log('readTV - Remote' +  JSON.stringify(myController)); 
+
+
+    ////
+*/
     this.bindCharacteristicEventsForInputs(myHarmonyAccessory);
 
     this.TVFoundAccessory = myHarmonyAccessory;